requirements.txt
streamlit==1.12.0
torch==2.8.0
torchvision==0.23.0
pandas==1.1.5
requests==2.32.5
Pillow==11.3.0
numpy==1.19.3
opencv-python==4.13.0.92